Inline highlight target into document highlight
