+++ lib/ruby_lsp/requests.rb
@@ -51,7 +51,6 @@ module RubyLsp
       autoload :SemanticTokenEncoder, "ruby_lsp/requests/support/semantic_token_encoder"
       autoload :Annotation, "ruby_lsp/requests/support/annotation"
       autoload :Sorbet, "ruby_lsp/requests/support/sorbet"
-      autoload :HighlightTarget, "ruby_lsp/requests/support/highlight_target"
       autoload :RailsDocumentClient, "ruby_lsp/requests/support/rails_document_client"
       autoload :Common, "ruby_lsp/requests/support/common"
       autoload :FormatterRunner, "ruby_lsp/requests/support/formatter_runner"
